requirements.txt
keras==2.3.1
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
Pillow==6.1
scikit_learn==0.22.1
seaborn==0.9.0
split_folders==0.3.1
torch==1.3.1
tqdm==4.41.1
